feat: add events stream and refactor stream organization
- Add new events stream processing pipeline with Kafka input and ClickHouse output
- Reorganize stream files into dedicated streams/ directory for better organization
- Update database migration processors with new naming conventions
- Add new eventstoslice processor for converting events to slice format
- Update signalstoslice processor with new naming convention
- Refactor Helm chart templates to dynamically mount stream files
- Update environment variables and configuration for unified database naming
- Update Makefile linting paths to reflect new stream organization
- Add new Kafka topic configuration for device events
