+++ charts/dps/files/streams/partial-cloudevents.yaml
